+++ src/field/field_testing.rs
@@ -136,7 +136,7 @@ pub fn run_binaryop_test_cases<F, BinaryOp, ExpectedOp>(
 macro_rules! test_arithmetic {
     ($field:ty) => {
         mod arithmetic {
-            use crate::{Field};
+            use crate::field::field::Field;
             use std::ops::{Add, Mul, Neg, Sub};
 
             // Can be 32 or 64; doesn't have to be computer's actual word
