+++ src/field/mod.rs
@@ -1,7 +1,6 @@
-pub(crate) mod crandall_field;
-pub(crate) mod field;
-pub(crate) mod field_search;
-pub(crate) mod fft;
+pub mod crandall_field;
+pub mod field;
+pub mod fft;
 pub(crate) mod cosets;
 
 #[cfg(test)]
